refactor(go-mod): Renamed `dependency` package to `process`.
