+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/go/gomod/process/GoRelationshipManager.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detectable.detectables.go.gomod.dependency;
+package com.synopsys.integration.detectable.detectables.go.gomod.process;
 
 import java.util.HashMap;
 import java.util.LinkedList;
